more dynamic action templates
